Add files via upload
Accenture Interview Question.

The problem statement: There is a table with fields employee_id, employee_name, and email_id. There are some duplicate records in this table. However, in case of duplicates, one of the records has email ID in lower case and the other has it in uppercase. Write a SQL to distinguish between the lower case email ID and the upper case email ID and keep the lowercase email ID record.